+++ Clienteditor/src/components/problemmodal_component.svelte
@@ -1,100 +1,188 @@
-<!--Modal-->
+<script>
+  import { apolloClient } from "../apolloClient.js";
+  import { getClient, query } from "svelte-apollo";
+  import TestModal from "../components/testmodal_component.svelte";
+  import ProblemModal from "../components/problemmodal_component.svelte";
+  import Navbar from "../components/navbar.svelte";
+  const client = getClient();
+  const Test = query(client, { query: apolloClient.allTests });
+  const Problem = query(client, { query: apolloClient.getProblems });
+</script>
 
-<div class="modalt opacity-0 pointer-events-none fixed w-full h-full top-0 left-0 flex items-center justify-center">
-    <div class="modal-overlayt absolute w-full h-full bg-gray-900 opacity-50"></div>
-    
-    <div class="modal-container bg-white w-11/12 md:max-w-md mx-auto rounded shadow-lg z-50 overflow-y-auto">
-    
-    <div class="modal-closet absolute top-0 right-0 cursor-pointer flex flex-col items-center mt-4 mr-4 text-white text-sm z-50">
-        <svg class="fill-current text-white" xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18">
-        <path d="M14.53 4.53l-1.06-1.06L9 7.94 4.53 3.47 3.47 4.53 7.94 9l-4.47 4.47 1.06 1.06L9 10.06l4.47 4.47 1.06-1.06L10.06 9z"></path>
-        </svg>
-        <span class="text-sm">(Esc)</span>
-    </div>
+<style>
+  .btn {
+    @apply font-bold py-2 px-4 rounded;
+  }
+  .btn-blue {
+    @apply bg-blue-500 text-white;
+  }
+  .btn-blue:hover {
+    @apply bg-blue-700;
+  }
+  button {
+    margin-left: 92%;
+  }
+  #blk {
+    margin-top: 6%;
+    margin-left: 10%;
+    margin-right: 6%;
+  }
+  #btn {
+    float: right;
+  }
+</style>
 
-    <!-- Add margin if you want to see some of the overlay behind the modal-->
-    <div class="modal-content py-4 text-left px-6">
-        <!--Title-->
-        <div class="flex justify-between items-center pb-3">
-        <p class="text-2xl font-bold">Add New Problem</p>
-        <div class="modal-closet cursor-pointer z-50">
-            <svg class="fill-current text-black" xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18">
-            <path d="M14.53 4.53l-1.06-1.06L9 7.94 4.53 3.47 3.47 4.53 7.94 9l-4.47 4.47 1.06 1.06L9 10.06l4.47 4.47 1.06-1.06L10.06 9z"></path>
-            </svg>
+<link
+  href="https://unpkg.com/tailwindcss@^1.0/dist/tailwind.min.css"
+  rel="stylesheet" />
+<body>
+
+  <Navbar />
+
+  <div id="blk" class="flex mb-4">
+    <div class="w-1/2 h-12">
+      <div class="max-w-lg rounded overflow-hidden shadow-lg">
+        <div class="px-6 py-4">
+          <div class="font-bold text-xl mb-2">Your Tests</div>
+          <ol>
+            {#await $Test}
+              Loading...
+            {:then result}
+              {#each result.data.allTests as test}
+                <li>
+                  <a href="http://localhost:5000/test/{test.id}">
+                    {test.testName}
+                  </a>
+                </li>
+              {/each}
+            {:catch err}
+              Error: {err}
+            {/await}
+          </ol>
         </div>
+        <div class="px-6 py-4" style="margin-right:5%;">
+          <button
+            class="modal-open bg-transparent border border-gray-500
+            hover:border-indigo-500 text-gray-500 hover:text-indigo-500
+            font-bold py-2 px-4 rounded-full">
+            Add Test
+          </button>
+          <TestModal />
         </div>
-
-        <!--Body-->
-        <form class="w-full max-w-md">
-<div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-    <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-full-name">
-        Problem Name
-    </label>
+      </div>
     </div>
-    <div class="md:w-2/3">
-    <input class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500" id="inline-full-name" type="text" value="Problem Name">
+    <div class="w-1/2 h-12">
+      <div class="max-w-lg rounded overflow-hidden shadow-lg">
+        <div class="px-6 py-4">
+          <div class="font-bold text-xl mb-2">Your Problems</div>
+          <ol>
+            {#await $Problem}
+              Loading...
+            {:then result}
+              {#each result.data.allProblems as prob}
+                <li>
+                  <a href="http://localhost:5000/problem/{prob.id}">
+                    {prob.problemName}
+                  </a>
+                </li>
+              {/each}
+            {:catch err}
+              Error: {err}
+            {/await}
+          </ol>
+        </div>
+        <div class="px-6 py-4" style="margin-right:10%;">
+          <button
+            class="modal-opent bg-transparent border border-gray-500
+            hover:border-indigo-500 text-gray-500 hover:text-indigo-500
+            font-bold py-2 px-4 rounded-full">
+            Add Problem
+          </button>
+          <ProblemModal />
+        </div>
+      </div>
     </div>
-</div>
-<div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-    <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-username">
-        Problem Description
-    </label>
-    </div>
-    <div class="md:w-2/3">
-    <textarea class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500" id="inline-username" type="text-area" placeholder="Problem Statement"></textarea>
-    </div>
-</div>
-<div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-    <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-username">
-        Test Cases
-    </label>
-    </div>
-    <div class="md:w-2/3">
-    <textarea class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500" id="inline-username" type="text-area" placeholder="eg : [1,3]"></textarea>
-    </div>
-</div>
-<div class="md:flex md:items-center mb-6">
-    <div class="md:w-1/3">
-    <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-username">
-        Output
-    </label>
-    </div>
-    <div class="md:w-2/3">
-    <textarea class="bg-gray-200 appearance-none border-2 border-gray-200 rounded w-full py-2 px-4 text-gray-700 leading-tight focus:outline-none focus:bg-white focus:border-purple-500" id="inline-username" type="text-area" placeholder="[4]"></textarea>
-    </div>
-</div>
-<div class="md:flex md:items-center mb-6">
-  <div class="md:w-1/3">
-    <label class="block text-gray-500 font-bold md:text-right mb-1 md:mb-0 pr-4" for="inline-username">
-        Difficulty Type
-    </label>
-    </div>
-  <div class="md:w-2/3">
-    <label class="inline-flex items-center">
-      <input type="radio" class="form-radio" name="difficultyType" value="easy">
-      <span class="ml-2 text-gray-500">Easy</span>
-    </label>
-    <label class="inline-flex items-center ml-6">
-      <input type="radio" class="form-radio" name="difficultyType" value="medium">
-      <span class="ml-2 text-gray-500">Medium</span>
-    </label>
-    <label class="inline-flex items-center ml-6">
-      <input type="radio" class="form-radio" name="difficultyType" value="hard">
-      <span class="ml-2 text-gray-500">Hard</span>
-    </label>
   </div>
-</div>
-</form>
 
-        <!--Footer-->
-        <div class="flex justify-end pt-2">
-        <button class="px-4 bg-transparent p-3 rounded-lg text-indigo-500 hover:bg-gray-100 hover:text-indigo-400 mr-2">Action</button>
-        <button class="modal-closet px-4 bg-indigo-500 p-3 rounded-lg text-white hover:bg-indigo-400">Close</button>
-        </div>
-        
-    </div>
-    </div>
-</div>
+  <script>
+    //Test Modal
+    var openmodal = document.querySelectorAll(".modal-open");
+    for (var i = 0; i < openmodal.length; i++) {
+      openmodal[i].addEventListener("click", function(event) {
+        event.preventDefault();
+        toggleModal();
+      });
+    }
+
+    const overlay = document.querySelector(".modal-overlay");
+    overlay.addEventListener("click", toggleModal);
+
+    var closemodal = document.querySelectorAll(".modal-close");
+    for (var i = 0; i < closemodal.length; i++) {
+      closemodal[i].addEventListener("click", toggleModal);
+    }
+
+    document.onkeydown = function(evt) {
+      evt = evt || window.event;
+      var isEscape = false;
+      if ("key" in evt) {
+        isEscape = evt.key === "Escape" || evt.key === "Esc";
+      } else {
+        isEscape = evt.keyCode === 27;
+      }
+      if (isEscape && document.body.classList.contains("modal-active")) {
+        toggleModal();
+      }
+    };
+
+    function toggleModal() {
+      const body = document.querySelector("body");
+      const modal = document.querySelector(".modal");
+      modal.classList.toggle("opacity-0");
+      modal.classList.toggle("pointer-events-none");
+      body.classList.toggle("modal-active");
+    }
+
+    // End of Test Modal
+
+    //Problem Modal
+    var openmodalt = document.querySelectorAll(".modal-opent");
+    for (var i = 0; i < openmodalt.length; i++) {
+      openmodalt[i].addEventListener("click", function(event) {
+        event.preventDefault();
+        toggleModalt();
+      });
+    }
+
+    const overlayt = document.querySelector(".modal-overlayt");
+    overlayt.addEventListener("click", toggleModalt);
+
+    var closemodalt = document.querySelectorAll(".modal-closet");
+    for (var i = 0; i < closemodalt.length; i++) {
+      closemodalt[i].addEventListener("click", toggleModalt);
+    }
+
+    document.onkeydown = function(evt) {
+      evt = evt || window.event;
+      var isEscape = false;
+      if ("key" in evt) {
+        isEscape = evt.key === "Escape" || evt.key === "Esc";
+      } else {
+        isEscape = evt.keyCode === 27;
+      }
+      if (isEscape && document.body.classList.contains("modal-active")) {
+        toggleModalt();
+      }
+    };
+
+    function toggleModalt() {
+      const body = document.querySelector("body");
+      const modal = document.querySelector(".modalt");
+      modal.classList.toggle("opacity-0");
+      modal.classList.toggle("pointer-events-none");
+      body.classList.toggle("modal-active");
+    }
+
+    // End of Problem Modal
+  </script>
+</body>
